% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casexwhen.R
\name{casexwhen}
\alias{casexwhen}
\title{Implements the case x when hack from SQL}
\usage{
casexwhen(input, map_in, map_out, lastelse = NULL, map_df = NULL)
}
\arguments{
\item{input}{A vector or scalar for the input values}

\item{map_in}{A vector or scalar for the input for the mapping}

\item{map_out}{A vector or scalar for the output for the mapping}

\item{lastelse}{A value to output if there are no matches, if \code{NULL} then no matches will output \code{NA}.}

\item{map_df}{A dataframe with the map_in as column 1 and map_out as column 2. When specified, takes priority over the specified input and output.}
}
\value{
A vector of the map_out values that correspond to the inputs.
}
\description{
Implements the case x when hack from SQL
}
\examples{
\dontrun{casexwhen(input=c("a", "c"), map_in=c("a", "b", "c", "d"), mapout=c(1,2,3,4))}
}
